v0.8 Sprint 1 (QA-013a): extract NativeTcpSocketBackend + NativeTcpSocketProbe from NativeTcpCarrier
First of a multi-PR decomposition track for NativeTcpCarrier
(1381 LOC, 11 class-level PMD suppressions, HEUR-061 carry-over from
v0.7 Sprint 8e). The carrier remains the TransportEngine implementation
and composes the new helpers; future PRs will extract ReactorLoop
(QA-013b) and client ingress/writer pumps (QA-013c).

Two new package-private classes:

  - NativeTcpSocketBackend (305 LOC, lifecycle owner)
    * SocketBackendMode enum (config value parsing + 4-source resolution)
    * SocketBackendSelection record (mode + arena + handles + armed flag)
    * Shared Arena ownership (deterministic close())
    * Validation latches (4 AtomicBoolean — server/client × attempted/successful)
    * Pkg-private getters used by NativeTcpCarrierBackendSelectionTest

  - NativeTcpSocketProbe (200 LOC, stateless probe helpers)
    * Server + client bootstrap probes (FFM downcalls)
    * IPv4 sockaddr serialization (BSD vs SysV layout)
    * Native socket() / bind() / connect() / listen() / close() wrappers
    * Best-effort WSACleanup() for Windows
    * Loopback constants + IS_WINDOWS_RUNTIME shared with Backend

The carrier delegates:

  - Constructor: this.backend = new NativeTcpSocketBackend()
  - start():     backend.validateServerSocketBootstrapOnce(allocator)
  - connect():   backend.validateClientSocketBackendOnce(allocator, host, port)
                 NativeTcpStream now receives backend.socketHandles()
  - close():     try { stop() } finally { backend.close() }
  - Tests:       carrier.requestedSocketBackend() / activeSocketBackend() /
                 isFfmSocketBackendArmed() / isServerSocketValidationSuccessful() /
                 isClientSocketValidationSuccessful() — all now thin proxies
                 to backend.* (preserves NativeTcpCarrierBackendSelectionTest
                 surface verbatim)

Public API surface unchanged. Sprint 7 transport non-blocking client
ingress refactor will be unaffected by this split.

Delta:
  NativeTcpCarrier 1381 → 1006 LOC (-375 / -27%)
  + NativeTcpSocketBackend 305 LOC
  + NativeTcpSocketProbe   200 LOC
  Carrier class-level suppressions: unchanged 11 (incremental track —
    QA-013b ReactorLoop + QA-013c client pumps will reduce them once
    further responsibility blocks land in their own files)

Test plan:
  - NativeTcpCarrierBackendSelectionTest 8/8 green (binary surface
    preserved via pkg-private proxies)
  - Full transport test suite 110/110 green
    (NativeTcpCarrierTckTest, NativeTcpStreamTckTest, MultiReactor{2,4}
    TckTest, CarrierPinningTckTest, IngressIntegrationTest,
    TransportStressTest — including TLS, plaintext, FFM-armed,
    NIO-fallback paths)
  - Full reactor verify SUCCESS (SPI + TCK + Core + Community Testkit +
    Community + Build Config + Parent + Root)
  - PMD 0 violations on community module
  - Checkstyle 0 violations on community module
